Crate wiwi_chain

Structs

Chain

OutputDebug

Traits

ChainConversions

Trait implemented on chains and their inner types, allowing you to get a reference to the inner type regardless of if the chain or the inner type is passed in

ChainConversionsOwned

ChainInner

Output

Safety

WithSelf

Functions

out_dbg

Tool for helping to debug Output trait usage in debug mode (if out is not written to, the function will panic)

Type Aliases

ArrayChain

ArrayMutChain

StringChain

StringMutChain

VecChain

VecMutChain