wiwi_chain

Struct Chain

pub struct Chain<T> { /* private fields */ }

Implementations

impl<T, const N: usize> Chain<[T; N]>

pub fn as_slice(self, cb: impl FnOnce(&[T])) -> Self

See documentation for [T; N]::as_slice for more details on the underlying function.

pub fn each_ref(self, cb: impl FnOnce([&T; N])) -> Self

See documentation for [T; N]::each_ref for more details on the underlying function.

pub fn each_mut(self, cb: impl FnOnce([&mut T; N])) -> Self

See documentation for [T; N]::each_mut for more details on the underlying function.

impl<'h, T, const N: usize> Chain<&'h mut [T; N]>

pub fn as_slice(self, cb: impl FnOnce(&[T])) -> Self

See documentation for [T; N]::as_slice for more details on the underlying function.

pub fn each_ref(self, cb: impl FnOnce([&T; N])) -> Self

See documentation for [T; N]::each_ref for more details on the underlying function.

pub fn each_mut(self, cb: impl FnOnce([&mut T; N])) -> Self

See documentation for [T; N]::each_mut for more details on the underlying function.

impl<T, const N: usize> Chain<[T; N]>

pub fn map<T2>(self, f: impl FnMut(T) -> T2) -> Chain<[T2; N]>

See documentation for [T; N]::map for more details on the underlying function.

impl Chain<String>

pub fn clear(self) -> Self

See documentation for String::clear for more details on the underlying function.

pub fn len(self, out: impl Output<usize>) -> Self

See documentation for String::len for more details on the underlying function.

pub fn push(self, ch: char) -> Self

See documentation for String::push for more details on the underlying function.

pub fn push_str(self, string: &str) -> Self

See documentation for String::push_str for more details on the underlying function.

pub fn reserve(self, additional: usize) -> Self

See documentation for String::reserve for more details on the underlying function.

pub fn reserve_exact(self, additional: usize) -> Self

See documentation for String::reserve_exact for more details on the underlying function.

pub fn retain(self, f: impl FnMut(char) -> bool) -> Self

See documentation for String::retain for more details on the underlying function.

impl<'h> Chain<&'h mut String>

pub fn clear(self) -> Self

See documentation for String::clear for more details on the underlying function.

pub fn len(self, out: impl Output<usize>) -> Self

See documentation for String::len for more details on the underlying function.

pub fn push(self, ch: char) -> Self

See documentation for String::push for more details on the underlying function.

pub fn push_str(self, string: &str) -> Self

See documentation for String::push_str for more details on the underlying function.

pub fn reserve(self, additional: usize) -> Self

See documentation for String::reserve for more details on the underlying function.

pub fn reserve_exact(self, additional: usize) -> Self

See documentation for String::reserve_exact for more details on the underlying function.

pub fn retain(self, f: impl FnMut(char) -> bool) -> Self

See documentation for String::retain for more details on the underlying function.

impl<T> Chain<Vec<T>>

pub unsafe fn align_to<U>(self, cb: impl FnOnce((&[T], &[U], &[T]))) -> Self

Safety

You must uphold safety invariants of [T]::align_to.

See documentation for [T]::align_to for more details on the underlying function.

pub unsafe fn align_to_mut<U>( self, cb: impl FnOnce((&mut [T], &mut [U], &mut [T])), ) -> Self

Safety

You must uphold safety invariants of [T]::align_to_mut.

See documentation for [T]::align_to_mut for more details on the underlying function.

pub fn append(self, other: &mut impl ChainConversions<Inner = Vec<T>>) -> Self

See documentation for Vec::append for more details on the underlying function.

pub fn binary_search( self, x: &T, out: impl Output<Result<usize, usize>>, ) -> Self
where T: Ord,

See documentation for [T]::binary_search for more details on the underlying function.

pub fn binary_search_by( self, f: impl FnMut(&T) -> Ordering, out: impl Output<Result<usize, usize>>, ) -> Self

See documentation for [T]::binary_search_by for more details on the underlying function.

pub fn binary_search_by_key<B>( self, b: &B, f: impl FnMut(&T) -> B, out: impl Output<Result<usize, usize>>, ) -> Self
where B: Ord,

See documentation for [T]::binary_search_by_key for more details on the underlying function.

pub fn capacity(self, out: impl Output<usize>) -> Self

See documentation for Vec::capacity for more details on the underlying function.

pub fn clear(self) -> Self

See documentation for Vec::clear for more details on the underlying function.

pub fn dedup(self) -> Self
where T: PartialEq,

See documentation for Vec::dedup for more details on the underlying function.

pub fn dedup_by(self, same_bucket: impl FnMut(&mut T, &mut T) -> bool) -> Self

See documentation for Vec::dedup_by for more details on the underlying function.

pub fn dedup_by_key<K>(self, key: impl FnMut(&mut T) -> K) -> Self
where K: PartialEq,

See documentation for Vec::dedup_by_key for more details on the underlying function.

pub fn fill(self, value: T) -> Self
where T: Clone,

See documentation for [T]::fill for more details on the underlying function.

pub fn fill_with(self, f: impl FnMut() -> T) -> Self

See documentation for [T]::fill_with for more details on the underlying function.

pub fn insert(self, index: usize, element: T) -> Self

See documentation for Vec::insert for more details on the underlying function.

pub fn is_empty(self, out: impl Output<bool>) -> Self

See documentation for Vec::is_empty for more details on the underlying function.

pub fn len(self, out: impl Output<usize>) -> Self

See documentation for Vec::len for more details on the underlying function.

pub fn pop(self, out: impl Output<Option<T>>) -> Self

See documentation for Vec::pop for more details on the underlying function.

pub fn push(self, value: T) -> Self

See documentation for Vec::push for more details on the underlying function.

pub fn remove(self, index: usize, out: impl Output<T>) -> Self

See documentation for Vec::remove for more details on the underlying function.

pub fn reserve(self, additional: usize) -> Self

See documentation for Vec::reserve for more details on the underlying function.

pub fn reserve_exact(self, additional: usize) -> Self

See documentation for Vec::reserve_exact for more details on the underlying function.

pub fn resize(self, new_len: usize, value: T) -> Self
where T: Clone,

See documentation for Vec::resize for more details on the underlying function.

pub fn resize_with(self, new_len: usize, f: impl FnMut() -> T) -> Self

See documentation for Vec::resize_with for more details on the underlying function.

pub fn retain(self, f: impl FnMut(&T) -> bool) -> Self

See documentation for Vec::retain for more details on the underlying function.

pub fn retain_mut(self, f: impl FnMut(&mut T) -> bool) -> Self

See documentation for Vec::retain_mut for more details on the underlying function.

pub unsafe fn set_len(self, new_len: usize) -> Self

Safety

You must uphold safety invariants of Vec::set_len.

See documentation for Vec::set_len for more details on the underlying function.

pub fn shrink_to(self, min_capacity: usize) -> Self

See documentation for Vec::shrink_to for more details on the underlying function.

pub fn shrink_to_fit(self) -> Self

See documentation for Vec::shrink_to_fit for more details on the underlying function.

pub fn sort(self) -> Self
where T: Ord,

See documentation for [T]::sort for more details on the underlying function.

pub fn sort_by(self, compare: impl FnMut(&T, &T) -> Ordering) -> Self

See documentation for [T]::sort_by for more details on the underlying function.

pub fn sort_by_key<K>(self, f: impl FnMut(&T) -> K) -> Self
where K: Ord,

See documentation for [T]::sort_by_key for more details on the underlying function.

pub fn sort_by_cached_key<K>(self, f: impl FnMut(&T) -> K) -> Self
where K: Ord,

See documentation for [T]::sort_by_cached_key for more details on the underlying function.

pub fn spare_capacity_mut(self, cb: impl FnOnce(&mut [MaybeUninit<T>])) -> Self

See documentation for Vec::spare_capacity_mut for more details on the underlying function.

impl<'h, T> Chain<&'h mut Vec<T>>

pub unsafe fn align_to<U>(self, cb: impl FnOnce((&[T], &[U], &[T]))) -> Self

Safety

You must uphold safety invariants of [T]::align_to.

See documentation for [T]::align_to for more details on the underlying function.

pub unsafe fn align_to_mut<U>( self, cb: impl FnOnce((&mut [T], &mut [U], &mut [T])), ) -> Self

Safety

You must uphold safety invariants of [T]::align_to_mut.

See documentation for [T]::align_to_mut for more details on the underlying function.

pub fn append(self, other: &mut impl ChainConversions<Inner = Vec<T>>) -> Self

See documentation for Vec::append for more details on the underlying function.

pub fn binary_search( self, x: &T, out: impl Output<Result<usize, usize>>, ) -> Self
where T: Ord,

See documentation for [T]::binary_search for more details on the underlying function.

pub fn binary_search_by( self, f: impl FnMut(&T) -> Ordering, out: impl Output<Result<usize, usize>>, ) -> Self

See documentation for [T]::binary_search_by for more details on the underlying function.

pub fn binary_search_by_key<B>( self, b: &B, f: impl FnMut(&T) -> B, out: impl Output<Result<usize, usize>>, ) -> Self
where B: Ord,

See documentation for [T]::binary_search_by_key for more details on the underlying function.

pub fn capacity(self, out: impl Output<usize>) -> Self

See documentation for Vec::capacity for more details on the underlying function.

pub fn clear(self) -> Self

See documentation for Vec::clear for more details on the underlying function.

pub fn dedup(self) -> Self
where T: PartialEq,

See documentation for Vec::dedup for more details on the underlying function.

pub fn dedup_by(self, same_bucket: impl FnMut(&mut T, &mut T) -> bool) -> Self

See documentation for Vec::dedup_by for more details on the underlying function.

pub fn dedup_by_key<K>(self, key: impl FnMut(&mut T) -> K) -> Self
where K: PartialEq,

See documentation for Vec::dedup_by_key for more details on the underlying function.

pub fn fill(self, value: T) -> Self
where T: Clone,

See documentation for [T]::fill for more details on the underlying function.

pub fn fill_with(self, f: impl FnMut() -> T) -> Self

See documentation for [T]::fill_with for more details on the underlying function.

pub fn insert(self, index: usize, element: T) -> Self

See documentation for Vec::insert for more details on the underlying function.

pub fn is_empty(self, out: impl Output<bool>) -> Self

See documentation for Vec::is_empty for more details on the underlying function.

pub fn len(self, out: impl Output<usize>) -> Self

See documentation for Vec::len for more details on the underlying function.

pub fn pop(self, out: impl Output<Option<T>>) -> Self

See documentation for Vec::pop for more details on the underlying function.

pub fn push(self, value: T) -> Self

See documentation for Vec::push for more details on the underlying function.

pub fn remove(self, index: usize, out: impl Output<T>) -> Self

See documentation for Vec::remove for more details on the underlying function.

pub fn reserve(self, additional: usize) -> Self

See documentation for Vec::reserve for more details on the underlying function.

pub fn reserve_exact(self, additional: usize) -> Self

See documentation for Vec::reserve_exact for more details on the underlying function.

pub fn resize(self, new_len: usize, value: T) -> Self
where T: Clone,

See documentation for Vec::resize for more details on the underlying function.

pub fn resize_with(self, new_len: usize, f: impl FnMut() -> T) -> Self

See documentation for Vec::resize_with for more details on the underlying function.

pub fn retain(self, f: impl FnMut(&T) -> bool) -> Self

See documentation for Vec::retain for more details on the underlying function.

pub fn retain_mut(self, f: impl FnMut(&mut T) -> bool) -> Self

See documentation for Vec::retain_mut for more details on the underlying function.

pub unsafe fn set_len(self, new_len: usize) -> Self

Safety

You must uphold safety invariants of Vec::set_len.

See documentation for Vec::set_len for more details on the underlying function.

pub fn shrink_to(self, min_capacity: usize) -> Self

See documentation for Vec::shrink_to for more details on the underlying function.

pub fn shrink_to_fit(self) -> Self

See documentation for Vec::shrink_to_fit for more details on the underlying function.

pub fn sort(self) -> Self
where T: Ord,

See documentation for [T]::sort for more details on the underlying function.

pub fn sort_by(self, compare: impl FnMut(&T, &T) -> Ordering) -> Self

See documentation for [T]::sort_by for more details on the underlying function.

pub fn sort_by_key<K>(self, f: impl FnMut(&T) -> K) -> Self
where K: Ord,

See documentation for [T]::sort_by_key for more details on the underlying function.

pub fn sort_by_cached_key<K>(self, f: impl FnMut(&T) -> K) -> Self
where K: Ord,

See documentation for [T]::sort_by_cached_key for more details on the underlying function.

pub fn spare_capacity_mut(self, cb: impl FnOnce(&mut [MaybeUninit<T>])) -> Self

See documentation for Vec::spare_capacity_mut for more details on the underlying function.

impl<T> Chain<T>

pub fn from_inner(inner: T) -> Self

pub fn into_inner(self) -> T

pub fn as_inner(&self) -> &T

pub fn as_inner_mut(&mut self) -> &mut T

pub fn with_inner(self, f: impl FnOnce(&mut T)) -> Self

Takes a closure that is called, passing in a reference to the inner value

This is useful for if there is something we have not implemented, or you otherwise need to borrow the inner struct for something, but are in the middle of some chain. It lets you do otherwise nonchainable operations inline with other chaining operations, so no need to break the chain c:

Examples

let chain = VecChain::<usize>::new();

// let's pretend `push` and `reserve` don't already have chainable versions...
let chain = chain
   .with_inner(|v| v.reserve(10))
   .with_inner(|v| v.push(1))
   .with_inner(|v| v.push(2));

assert!(chain.as_inner().len() == 2);
assert!(chain.as_inner().capacity() >= 10);

impl<T> Chain<&T>
where T: Clone,

pub fn cloned(&self) -> Chain<T>

impl<T> Chain<&mut T>
where T: Clone,

pub fn cloned(&self) -> Chain<T>

impl<T> Chain<&T>
where T: Copy,

pub fn copied(&self) -> Chain<T>

impl<T> Chain<&mut T>
where T: Copy,

pub fn copied(&self) -> Chain<T>

impl<'h, T> Chain<&'h &'h T>

pub fn flatten(self) -> Chain<&'h T>

impl<'h, T> Chain<&'h mut &'h mut T>

pub fn flatten(self) -> Chain<&'h mut T>